Fix annoying, rare bug that caused context stack to become unbalanced.
The bug occurred when the name of an unnamed TableSpec (nil) was
pushed onto the context stack, thus effectively leaving the stack
unchanged. The corresponding pop however removed an entry.

Added an (optional) simple stack checking mechanism by returning the
depth when pushing and checking that when popping.

Also improved the log output by adding indentation.